chore: add separate docker-compose file for Cypress container
the commented out sections were part of an attempt to have the app run
in a container as well that the Cypress container can talk to.
I couldn't get it to work, so for now, the Cypress container is
dependent on the app running and accessible from your host machine.